rest for users

+++ soplib-restws/src/sop/library/model/builder/UserModelBuilderImpl.java
@@ -6,18 +6,11 @@ import java.util.List;
 import sop.library.dal.dao.UserDao;
 import sop.library.dal.dao.UserDaoImpl;
 import sop.library.dal.entities.usermanagement.UserEntity;
+import sop.library.exceptions.ResourceNotFoundException;
 import sop.library.model.User;
 import sop.library.model.mapper.UserMapper;
 
 public class UserModelBuilderImpl implements UserModelBuilder {
-	
-	public Long save(User userDto) {
-		UserDao userDao = new UserDaoImpl();
-		UserEntity user = UserMapper.buildFromDto(userDto);
-		userDao.saveOrUpdate(user);
-		Long savedUserId = user.getId();
-		return savedUserId;
-	}
 	
 	public List<User> get() {
 		List<User> users = new ArrayList<User>();
@@ -28,26 +21,34 @@ public class UserModelBuilderImpl implements UserModelBuilder {
 		return users;
 	}
 	
-	public User get(Long id) {
+	public User get(Long id) throws ResourceNotFoundException {
 		UserDao userDao = new UserDaoImpl();
 		UserEntity user = userDao.find(id);
 		if(user == null)
-			return null;
+			throw new ResourceNotFoundException();
 		User userDto = UserMapper.buildFromEntity(user);
 		return userDto;
 	}
-
-	public void update(Long id, User userDto) {
+	
+	public Long save(User userDto) {
+		UserDao userDao = new UserDaoImpl();
+		UserEntity user = UserMapper.buildFromDto(userDto);
+		userDao.saveOrUpdate(user);
+		Long savedUserId = user.getId();
+		return savedUserId;
+	}
+	
+	public void update(Long id, User userDto) throws ResourceNotFoundException {
 		UserDao userDao = new UserDaoImpl();
 		UserEntity user = UserMapper.mapUser(id, userDto);
 		userDao.saveOrUpdate(user);
 	}
 	
-	public void delete(Long id) {
+	public void delete(Long id) throws ResourceNotFoundException {
 		UserDao userDao = new UserDaoImpl();
 		UserEntity user = userDao.find(id);
+		if(user == null)
+			throw new ResourceNotFoundException();
 		userDao.delete(user);
 	}
-
-	
 }
